Rename Path files
Path moved down to CoreCLR, the code here is only for CoreRT. Renaming
and combining the code to reflect the usage.
