+++ src/System.Runtime.Extensions/src/System/IO/PathHelper.Uap.cs
